chore(application): reorganize server routers
